requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
amqp
asgiref
asttokens
async-timeout
attrs
autopep8
backcall
beautifulsoup4
billiard
celery
certifi
cffi
chardet
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
decorator
defusedxml
diff-match-patch
Django
django-allauth
django-ckeditor-5
django-crispy-forms
django-csp
django-debug-toolbar
django-import-export
django-js-asset
django-mptt
django-recaptcha3
django-tinymce
django-user-agents
et_xmlfile
exceptiongroup
executing
frozenlist
gunicorn
h11
idna
ipython
ipython-genutils
jdcal
jedi
kombu
lxml
MarkupPy
matplotlib-inline
multidict
oauthlib
odfpy
openpyxl
packaging
parso
pep8
pexpect
pickleshare
pillow
prompt_toolkit
propcache
psycopg2
ptyprocess
pure_eval
pycodestyle
pycparser
Pygments
PyJWT
python-dateutil
python-decouple
python-dotenv
python3-openid
pytils
pytz
PyYAML
redis
requests
requests-oauthlib
six
soupsieve
sqlparse
stack-data
stripe
tablib
toml
tomli
traitlets
transliterate
typing_extensions
tzdata
ua-parser
ua-parser-builtins
urllib3
user-agents
uvicorn
vine
wcwidth
xlrd
xlwt
yarl
